+++ pkgs/checks/lib/src/extensions/function.dart
@@ -1,7 +1,10 @@
 // Copyright (c) 2022, the Dart project authors.  Please see the AUTHORS file
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
+import 'dart:async';
 import 'dart:convert';
+
+import 'package:meta/meta.dart' as meta;
 
 import '../../context.dart';
 
@@ -71,4 +74,83 @@ extension FunctionChecks<T> on Subject<T Function()> {
     },
     nestedCondition: that,
   );
+
+  /// Expects that the function prints text when called.
+  ///
+  /// Intercepts calls to [print] while the function is executed and returns
+  /// a [Subject] to check expectations on the captured printed output.
+  ///
+  /// If the function throws synchronously, this expectation will fail.
+  ///
+  /// If this function is async and returns a [Future], this expectation will
+  /// fail. Use [printsAsync] instead.
+  Subject<String> prints([Condition<String>? that]) {
+    return context.nest<String>(() => ['prints'], (actual) {
+      final buffer = StringBuffer();
+      try {
+        final result = runZoned(
+          actual,
+          zoneSpecification: ZoneSpecification(
+            print: (_, _, _, line) {
+              buffer.writeln(line);
+            },
+          ),
+        );
+        if (result is Future) {
+          return Extracted.rejection(
+            actual: ['a function that returned a Future'],
+            which: [
+              'returned a Future; use printsAsync to test asynchronous functions',
+            ],
+          );
+        }
+        return Extracted.value(buffer.toString());
+      } catch (e, st) {
+        return Extracted.rejection(
+          actual: ['a function that throws'],
+          which: [
+            ...prefixFirst('threw ', postfixLast(' at:', literal(e))),
+            ...indent(LineSplitter.split(st.toString())),
+          ],
+        );
+      }
+    }, nestedCondition: that);
+  }
+
+  /// Expects that the function prints text when called and completed.
+  ///
+  /// Intercepts calls to [print] while the function is executed and waits
+  /// for the returned [Future] to complete before checking expectations
+  /// on the captured printed output.
+  ///
+  /// If the function throws synchronously or returns a [Future] that completes
+  /// with an error, this expectation will fail.
+  @meta.awaitNotRequired
+  Future<Subject<String>> printsAsync([Condition<String>? printCondition]) {
+    return context.nestAsync<String>(() => ['prints'], (actual) async {
+      final buffer = StringBuffer();
+      try {
+        final result = runZoned(
+          actual,
+          zoneSpecification: ZoneSpecification(
+            print: (_, _, _, line) {
+              buffer.writeln(line);
+            },
+          ),
+        );
+        if (result is Future) {
+          await result;
+        }
+        return Extracted.value(buffer.toString());
+      } catch (e, st) {
+        return Extracted.rejection(
+          actual: ['a function that throws'],
+          which: [
+            ...prefixFirst('threw ', postfixLast(' at:', literal(e))),
+            ...indent(LineSplitter.split(st.toString())),
+          ],
+        );
+      }
+    }, printCondition);
+  }
 }