+++ pkgs/checks/CHANGELOG.md
@@ -39,6 +39,8 @@
   `(await check(foo).completes()).equals(42);`
 - Added `FutureSubjectExtension.which` to allow chaining expectations on a
   `Future<Subject>`.
+- Add `prints()` and `printsAsync()` check extensions on `Subject<T Function()>`
+  to check expectations against captured printed output.
 
 ## 0.3.2